Rename voter app to constituencies
It's clear that I have created a constituency app rather than one for voters so I'm renaming it.

WARNING: deploying this will break an existing Django installation without additional work! There are some database changes that need to be made in addition.

Exporting the data, dropping the DB, recreating it, running manage.py migrate, then re-importing the data should work.

Here is some advice (which I have not tested) on altering the DB rather than recreating it:

http://stackoverflow.com/questions/8408046/how-to-change-the-name-of-a-django-app
