+++ q_and_a/apps/constituencies/urls.py
@@ -2,5 +2,5 @@
 from django.conf.urls import patterns, url
 
 urlpatterns = patterns('',
-    url(r'^(.+)/$', 'voters.views.ConstituencyView', name='constituency'),
+    url(r'^(.+)/$', 'constituencies.views.ConstituencyView', name='constituency'),
 )
