+++ q_and_a/urls.py
@@ -10,11 +10,11 @@ from django.contrib import admin
 admin.autodiscover()
 
 urlpatterns = patterns('',
-    url(r'^$', 'voters.views.HomePageView', name='home'),
+    url(r'^$', 'constituencies.views.HomePageView', name='home'),
     # Uncomment the next line to enable the admin:
     url(r'^admin/', include(admin.site.urls)),
     url(r'^organisations', include('organisations.urls')),
     url(r'^candidates', include('candidates.urls')),
     url(r'^api/v1/question/(\d+)/?$', 'questions.views.api'),
-    url(r'^constituencies/', include('voters.urls')),
+    url(r'^constituencies/', include('constituencies.urls')),
 ) + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
